refactor: page.tsx에서 fetch 요청
