+++ src/app/my/sentence/components/MySentenceList/index.module.scss
